\name{DESP-internal}
% functions documented as internal
\alias{DESP_MST_Phi}
\alias{DESP_SPT_Phi}
\alias{DESP_SPT_Phi2}
\alias{scs_SOCP_solve}
\alias{scs_sqR_Lasso}
\alias{scs_SRL_B}
\title{Internal DESP Functions}
\description{
  Internal functions used in DESP. 
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
