% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESP.R
\name{DESP}
\alias{DESP}
\title{DESP: Demixing Cell State Profiles from Dynamic Bulk Molecular
Measurements}
\arguments{
\item{bulk}{Bulk omics measurements. Feature-by-sample numeric matrix.
Features can correspond to genes (e.g. RNA-Seq data) or proteins (e.g.
proteomics data). Note: \emph{DESP} does not perform any pre-processing of the
data.}

\item{proportions}{Cell state proportion matrix. Sample-by-cell state
numeric matrix indicating either the number of cells or relative proportion
of each cell state in each sample. These cell state proportions can be
derived from techniques such as scRNA-Seq or FACS.}

\item{lambda}{Ridge regression normalization parameter.
Default value is \code{1e-7} and should be suitable for most applications.}

\item{beta}{Cell state similarity normalization parameter.
Default value is \code{1e-4} and should be suitable for most applications.}

\item{similarities}{(Optional) Cell state similarity matrix. State-by-state numeric
matrix containing the similarity of the cell states to each other. Any
similarity metric can be used (e.g. Pearson correlation) as long as the
larger values indicate higher similarity (e.g. distances should be
converted to similarities). An example is the correlation of per-state
averaged scRNA-Seq gene expression values.}
}
\value{
\emph{DESP}-predicted cell state profiles. Feature-by-cell state
numeric matrix containing the predicted feature levels for each of the
pre-defined cell states, effectively demixing the measured bulk 'omics
matrix into the underlying cell state 'omics profiles.
}
\description{
Predict cell state \emph{'omics} profiles given dynamic bulk
\emph{'omics} measurements and corresponding cell state proportions across
samples.
}
\details{
Note: samples should be consistently named/ordered between the two datasets
and not contain missing values. We recommend reversing log transformations
before applying \emph{DESP} to align with its additive model ('bulk as sum of
single cells'), followed by log transformation of resultant cell
state-level predictions for downstream analyses as needed.
}
\examples{
#read in example bulk omics data
data(bulk_EMT_proteome)

#read in example cell state proportions
data(EMT_proportions)

#DESP demixing to predict cell state profiles
prediction <- DESP(bulk_EMT_proteome, EMT_proportions)

}
