% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeDESP.R
\name{visualizeDESP}
\alias{visualizeDESP}
\title{Visualize DESP demixing of a given feature}
\arguments{
\item{bulk}{Bulk omics measurements. Feature-by-sample numeric matrix.
Features can correspond to genes (e.g. RNA-Seq data) or proteins (e.g.
proteomics data). Note: \emph{DESP} does not perform any pre-processing of the
data.}

\item{prediction}{\emph{DESP}-predicted cell state profiles. Feature-by-cell
state numeric matrix containing the \emph{DESP}-predicted feature levels for
each of the pre-defined cell states.}

\item{proportions}{Cell state proportion matrix. Sample-by-cell state
numeric matrix indicating either the number of cells or relative proportion
of each cell state in each sample. These cell state proportions can be
derived from techniques such as scRNA-Seq or FACS.}

\item{feature}{Feature to visualize the associated \emph{DESP} demixing.
Feature must be present in both input matrices
(\code{bulk} & \code{prediction}).}

\item{highlightSample}{(Optional) Sample to highlight in plot.
Must correspond to one of the column names in \code{bulk} input matrix.}
}
\value{
Plot visualizing \emph{DESP}'s predicted cell state
contributions to bulk measurements for a given feature (e.g. gene/protein).
Plot is created using the ggplot2, ggsankey, and ggalluvial packages.
}
\description{
Create plot visualizing \emph{DESP}'s predicted cell state
contributions to bulk measurements for a given feature (e.g. gene/protein).

Blocks on the left side of the plot correspond to input bulk measurements
of given feature, while blocks on the right correspond to the predicted
cell state levels. Thickness of lines connecting blocks correspond to
relative contribution of cell state (right) to bulk sample (left), with
labels for contributions greater than 15\%.

Note: This function requires the output of the \code{DESP()} function.
Make sure to run \emph{DESP} on your data first!
}
\examples{
#read in example bulk omics data
data(bulk_EMT_proteome)

#read in example cell state proportions
data(EMT_proportions)

#DESP demixing to predict cell state profiles
prediction <- DESP(bulk_EMT_proteome, EMT_proportions)

#visualize results for example protein 'AAK1'
visualizeDESP(bulk = bulk_EMT_proteome,
prediction = prediction,
proportions = EMT_proportions,
feature = "AAK1")

}
